# wfc/canvas/static/scripts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NodeRunState": {
        "properties": {
          "cache_hit": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Hit"
          },
          "cache_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Key"
          },
          "cancelled_due_to_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled Due To Run Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "error_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Run Id"
          },
          "error_sample": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Sample"
          },
          "original_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Run Id"
          },
          "push_failed_count": {
            "default": 0,
            "title": "Push Failed Count",
            "type": "integer"
          },
          "push_pending_count": {
            "default": 0,
            "title": "Push Pending Count",
            "type": "integer"
          },
          "push_state": {
            "default": "deferred",
            "title": "Push State",
            "type": "string"
          },
          "run_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Ids"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tally": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tally"
          },
          "upstream_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Node Id"
          },
          "upstream_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Run Id"
          }
        },
        "required": [
          "status"
        ],
        "title": "NodeRunState",
        "type": "object"
      },
      "PipelineInput": {
        "properties": {
          "explicit_combos": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explicit Combos"
          },
          "keep_going": {
            "default": false,
            "title": "Keep Going",
            "type": "boolean"
          },
          "links": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PipelineLink"
            },
            "title": "Links",
            "type": "array"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/PipelineNode"
            },
            "title": "Nodes",
            "type": "array"
          },
          "param_sets": {
            "anyOf": [
              {
                "additionalProperties": {
                  "additionalProperties": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "object"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Param Sets"
          },
          "samples": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Samples",
            "type": "array"
          },
          "variables": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variables"
          }
        },
        "required": [
          "nodes"
        ],
        "title": "PipelineInput",
        "type": "object"
      },
      "PipelineLink": {
        "properties": {
          "source": {
            "title": "Source",
            "type": "string"
          },
          "sourceHandle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcehandle"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "targetHandle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targethandle"
          }
        },
        "required": [
          "source",
          "target"
        ],
        "title": "PipelineLink",
        "type": "object"
      },
      "PipelineNode": {
        "properties": {
          "fan_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fan Mode"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "method": {
            "default": "",
            "title": "Method",
            "type": "string"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Module"
          },
          "params": {
            "additionalProperties": true,
            "default": {},
            "title": "Params",
            "type": "object"
          },
          "position": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "samples": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Samples"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "method",
            "title": "Type"
          }
        },
        "required": [
          "id"
        ],
        "title": "PipelineNode",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkflowStatusResponse": {
        "properties": {
          "error": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "log": {
            "title": "Log",
            "type": "string"
          },
          "node_states": {
            "additionalProperties": {
              "$ref": "#/components/schemas/NodeRunState"
            },
            "title": "Node States",
            "type": "object"
          },
          "overall_status": {
            "title": "Overall Status",
            "type": "string"
          },
          "thread_alive": {
            "title": "Thread Alive",
            "type": "boolean"
          }
        },
        "required": [
          "job_id",
          "overall_status",
          "node_states",
          "thread_alive",
          "log"
        ],
        "title": "WorkflowStatusResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Workflow Canvas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/workflow/cancel/{job_id}": {
      "post": {
        "description": "Cancel an in-flight pipeline.\n\nTerminates the live Snakemake subprocess (and its descendants) and\nflips any ``running`` rows for this pipeline to ``cancelled`` with\n``error_message=\"Cancelled by user\"``.  Idempotent: cancelling a\npipeline whose process has already exited returns 200 with a no-op\nindication.  Returns 404 for unknown job_ids.",
        "operationId": "cancel_workflow_api_workflow_cancel__job_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Workflow"
      }
    },
    "/api/workflow/run": {
      "post": {
        "description": "Submit a pipeline for execution via Snakemake.\n\nEnriches the PipelineJSON with script paths and slot_outputs from the DB,\nwrites it to a temp file, and spawns a background thread running\n``run_pipeline()``. Returns the pipeline_id as the job_id immediately.\n\nRejects empty pipelines with HTTP 400.  Before spawning the run thread it\npre-flights Docker and git: a not-ready\nenvironment is rejected with HTTP 409 carrying a kind-tagged\n``{kind, message, hint}`` payload (the same shape the frontend renders for\npre-run errors), so no orphan run is started.  Multiple pipelines may run\nconcurrently \u2014 each gets its own pipeline_id-scoped workspace.",
        "operationId": "run_workflow_api_workflow_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Workflow"
      }
    },
    "/api/workflow/status/{job_id}": {
      "get": {
        "description": "Return per-node and overall status for a running or completed pipeline.\n\nQueries the ``runs`` table by ``pipeline_id`` and derives overall status.\nAlso returns whether the background thread is still alive and any captured\nlog output.\n\nArgs:\n    job_id: The pipeline_id returned by the run endpoint.\n\nReturns:\n    JSON with ``job_id``, ``overall_status``, ``node_states``, ``thread_alive``,\n    ``log``, and ``error`` fields.",
        "operationId": "get_workflow_status_api_workflow_status__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow Status"
      }
    },
...
  }
}